reorganize functions for checks
